Support prefetching images based on last revision.
In my test of Governor generals of canada page that has 85 images,
This brings the time to render down to 1.05 seconds (A 99.8%
improvement of vanilla ForeignAPIRepo's 17 minutes). Probably a lot
of that time is not even the file stuff anymore but other rendering
stuff.

An interesting note - we had to enable CURLOPT_PIPEWAIT to actually
use HTTP/2 pipelining. Core might want to make a similar change.

Change-Id: I9cf3e7a3b7055e80e5f77fd4449559bd1236e55f