Cleanup i18n a little bit
This duplicates some code, because I don't want to reuse the
$repo->getDescription(), because its fallback is "a shared repository",
but the correct grammar in english would be "the shared repository".
It seems like even in core that message is used both in places where
"the" makes sense, and palces where "a" makes sense. I imagine it
only gets worse in other languages.

Bump version to 1.1

Change-Id: I6fc998405ec8ab4193a0a9336218427329086609