Use normalized titles from API of foreign repo
On wikis with $wgCapitalLinks disabled, downstream consumers of the
File class may attempt to build an image URL under the assumption
that the title is also lowercase on the foreign repo. This means
the wrong title and hash are used, and an invalid URL will be
generated. This is the case when using Commons as a foreign repo
and when a gallery tag is used.

This ensures that normalized (uppercase-first converted) filenames
provided by the foreign repo's API are used instead. This means the
correct title will be used locally even if it uses non-capitalized
links.

Bug: T391750
Change-Id: I29957535077adbc65648cf652024e01a3a55518c
(cherry picked from commit 1af3264505b242b92bdee3f6c5c656409cc1b684)